1.7.3.0 (unstable/experimental)
* sync various pieces of code with Incentive PyMOL
* faster iterate/alter implementation ported from Incentive PyMOL
* experimental mmCIF write support (atoms only)
* partial multi-letter chain support
* super: use guide instead of CA, enables nuc acid alignment
* fix movie panel not shown until resize
* eliminate some deprecated parsing modes
* python: convert some files to absolute_import
* --help and --version
* dynamic_measures refactoring, fixes duplicated IDs bug
* get_type returns object:alignment and object:ramp
* new "command" Wizard ported from Incentive PyMOL
* remove obsolete contrib/sglite and contrib/modules

+++ layer0/Triangle.c
@@ -816,7 +816,7 @@ static int TriangleBuildObvious(TriangleSurfaceRec * II, int i1, int i2, float *
                   minDist2 = dif2;
                   maxDot = dot;
                   i0 = j;
-                } else if((dif2 / minDist2) < (float) pow(2 * (dot / maxDot), 2.0)) {
+                } else if((dif2 / minDist2) < powf(2 * (dot / maxDot), 2.0f)) {
                   minDist2 = dif2;
                   maxDot = dot;
                   i0 = j;
@@ -1021,7 +1021,7 @@ static int TriangleBuildSecondPass(TriangleSurfaceRec * II, int i1, int i2, floa
                   minDist2 = dif2;
                   maxDot = dot;
                   i0 = j;
-                } else if((dif2 / minDist2) < pow(2 * (dot / maxDot), 2.0)) {
+                } else if((dif2 / minDist2) < powf(2 * (dot / maxDot), 2.0f)) {
                   maxDot = dot;
                   minDist2 = dif2;
                   i0 = j;
